Zne pec test uniformity (#676)
* Qiskit tests combined

Added qiskit tests to test_zne

* Combined PyQuil

Added PyQuil tests to test_zne.py

* CONTRIBUTING updated

Mentioned the requirement for Qiskit and pyQuil to tests

* Formatting

Fixed formatting

* Naming cleanup

* Reintroduced Pyquil ZNE tests

* Removed pyQuil tests from zne folder

* Updated makefile

* Updated documentation
