+++ CONTRIBUTING.md
@@ -48,27 +48,33 @@ development of the library, from your local clone of the fork, run
 
 ### Adding tests
 If you add new features to a function or class, it is required to add tests for such object. Mitiq uses a nested structure for packaging tests in directories named `tests` at the same level of each module.
+The only except to this is that any tests requiring a QVM should be placed in the mitiq_pyquil/tests folder.
 
 ### Updating the documentation
 Follow these [instructions for contributing to the documentation](https://mitiq.readthedocs.io/en/latest/contributing_docs.html) which include guidelines about updating the API-doc list of modules and writing examples in the users guide.
 
-### Checking local tests
+### Running local tests
 
+After making changes, please ensure your changes still pass all the existing tests.
 You can check that tests run with `pytest`. The [Makefile][makefile] contains
 some commands for running different collections of tests for the repository.
 
-To run just the tests contained in `mitiq/tests` and `mitiq/benchmarks/tests` run
+To only run tests that do not require a pyQuil QVM running, run
 
 ```bash
 (myenv) make test
 ```
 
-To run the tests for the pyQuil and Qiskit plugins (which of course require for
-pyQuil and Qiskit to be installed) run
+To run the tests for the pyQuil plugins, run
 
 ```bash
 (myenv) make test-pyquil
-(myenv) make test-qiskit
+```
+
+To run all tests, run
+
+```bash
+(myenv) make test-all
 ```
 
 *NOTE*: For the pyQuil tests to run, you will need to have QVM & quilc servers
@@ -79,15 +85,13 @@ docker run --rm -idt -p 5000:5000 rigetti/qvm -S
 docker run --rm -idt -p 5555:5555 rigetti/quilc -R
 ```
 
-You can also check that all tests run also in the documentation examples and
+Please also remember to check that all tests run also in the documentation examples and
 docstrings with
 
 ```bash
-(myenv) make docs
+(myenv) make doctest
 ```
 
-If you add new `/tests` directories, you will need to update the `Makefile`
-so that they will be included as part of continuous integration.
 
 ### Style Guidelines
