+++ mitiq/zne/tests/test_zne.py
@@ -16,16 +16,34 @@
 """Unit tests for zero-noise extrapolation."""
 import numpy as np
 import pytest
-
 import cirq
+import qiskit
 
+from mitiq.zne import (
+    inference,
+    scaling,
+    execute_with_zne,
+    mitigate_executor,
+    zne_decorator,
+)
 from mitiq.zne.inference import LinearFactory, RichardsonFactory
 from mitiq.zne.scaling import (
     fold_gates_from_left,
     fold_gates_from_right,
     fold_gates_at_random,
 )
-from mitiq.zne import execute_with_zne, mitigate_executor, zne_decorator
+from mitiq._typing import QPROGRAM
+from mitiq.benchmarks.randomized_benchmarking import generate_rb_circuits
+
+from mitiq.mitiq_qiskit import (
+    execute_with_shots_and_noise,
+    initialized_depolarizing_noise,
+)
+
+BASE_NOISE = 0.007
+TEST_DEPTH = 30
+ONE_QUBIT_GS_PROJECTOR = np.array([[1, 0], [0, 0]])
+QASM_SIMULATOR = qiskit.Aer.get_backend("qasm_simulator")
 
 npX = np.array([[0, 1], [1, 0]])
 """Defines the sigma_x Pauli matrix in SU(2) algebra as a (2,2) `np.array`."""
@@ -143,3 +161,206 @@ def test_doc_is_preserved():
         return 0
 
     assert second_executor.__doc__ == first_executor.__doc__
+
+
+def qiskit_measure(circuit, qid) -> qiskit.QuantumCircuit:
+    """Helper function to measure one qubit."""
+    # Ensure that we have a classical register of enough size available
+    if len(circuit.clbits) == 0:
+        reg = qiskit.ClassicalRegister(qid + 1, "cbits")
+        circuit.add_register(reg)
+    circuit.measure(0, qid)
+    return circuit
+
+
+def qiskit_executor(qp: QPROGRAM, shots: int = 500) -> float:
+    # initialize a qiskit noise model
+    expectation = execute_with_shots_and_noise(
+        qp,
+        shots=shots,
+        obs=ONE_QUBIT_GS_PROJECTOR,
+        noise_model=initialized_depolarizing_noise(BASE_NOISE),
+        seed=1,
+    )
+    return expectation
+
+
+def get_counts(circuit: qiskit.QuantumCircuit):
+    return (
+        qiskit.execute(
+            circuit, qiskit.Aer.get_backend("qasm_simulator"), shots=100
+        )
+        .result()
+        .get_counts()
+    )
+
+
+def test_qiskit_execute_with_zne():
+    true_zne_value = 1.0
+
+    circuit = qiskit_measure(
+        *generate_rb_circuits(
+            n_qubits=1,
+            num_cliffords=TEST_DEPTH,
+            trials=1,
+            return_type="qiskit",
+        ),
+        0,
+    )
+    base = qiskit_executor(circuit)
+    zne_value = execute_with_zne(circuit, qiskit_executor)
+
+    assert abs(true_zne_value - zne_value) < abs(true_zne_value - base)
+
+
+@zne_decorator()
+def qiskit_decorated_executor(qp: QPROGRAM) -> float:
+    return qiskit_executor(qp)
+
+
+def test_qiskit_mitigate_executor():
+    true_zne_value = 1.0
+
+    circuit = qiskit_measure(
+        *generate_rb_circuits(
+            n_qubits=1,
+            num_cliffords=TEST_DEPTH,
+            trials=1,
+            return_type="qiskit",
+        ),
+        0,
+    )
+    base = qiskit_executor(circuit)
+
+    mitigated_executor = mitigate_executor(qiskit_executor)
+    zne_value = mitigated_executor(circuit)
+    assert abs(true_zne_value - zne_value) < abs(true_zne_value - base)
+
+
+def test_qiskit_zne_decorator():
+    true_zne_value = 1.0
+
+    circuit = qiskit_measure(
+        *generate_rb_circuits(
+            n_qubits=1,
+            num_cliffords=TEST_DEPTH,
+            trials=1,
+            return_type="qiskit",
+        ),
+        0,
+    )
+    base = qiskit_executor(circuit)
+
+    zne_value = qiskit_decorated_executor(circuit)
+    assert abs(true_zne_value - zne_value) < abs(true_zne_value - base)
+
+
+def test_qiskit_run_factory_with_number_of_shots():
+    true_zne_value = 1.0
+
+    scale_factors = [1.0, 2.0, 3.0]
+    shot_list = [1_000, 2_000, 3_000]
+
+    fac = inference.ExpFactory(
+        scale_factors=scale_factors, shot_list=shot_list
+    )
+
+    circuit = qiskit_measure(
+        *generate_rb_circuits(
+            n_qubits=1,
+            num_cliffords=TEST_DEPTH,
+            trials=1,
+            return_type="qiskit",
+        ),
+        0,
+    )
+    base = qiskit_executor(circuit)
+    zne_value = fac.run(
+        circuit, qiskit_executor, scale_noise=scaling.fold_gates_at_random,
+    ).reduce()
+
+    assert abs(true_zne_value - zne_value) < abs(true_zne_value - base)
+
+    for i in range(len(fac._instack)):
+        assert fac._instack[i] == {
+            "scale_factor": scale_factors[i],
+            "shots": shot_list[i],
+        }
+
+
+def test_qiskit_mitigate_executor_with_shot_list():
+    true_zne_value = 1.0
+
+    scale_factors = [1.0, 2.0, 3.0]
+    shot_list = [1_000, 2_000, 3_000]
+
+    fac = inference.ExpFactory(
+        scale_factors=scale_factors, shot_list=shot_list
+    )
+    mitigated_executor = mitigate_executor(qiskit_executor, fac)
+
+    circuit = qiskit_measure(
+        *generate_rb_circuits(
+            n_qubits=1,
+            num_cliffords=TEST_DEPTH,
+            trials=1,
+            return_type="qiskit",
+        ),
+        0,
+    )
+    base = qiskit_executor(circuit)
+    zne_value = mitigated_executor(circuit)
+
+    assert abs(true_zne_value - zne_value) < abs(true_zne_value - base)
+
+    for i in range(len(fac._instack)):
+        assert fac._instack[i] == {
+            "scale_factor": scale_factors[i],
+            "shots": shot_list[i],
+        }
+
+
+@pytest.mark.parametrize("order", [(0, 1), (1, 0), (0, 1, 2), (1, 2, 0)])
+def test_qiskit_measurement_order_is_preserved_single_register(order):
+    """Tests measurement order is preserved when folding, i.e., the dictionary
+    of counts is the same as the original circuit on a noiseless simulator.
+    """
+    qreg, creg = (
+        qiskit.QuantumRegister(len(order)),
+        qiskit.ClassicalRegister(len(order)),
+    )
+    circuit = qiskit.QuantumCircuit(qreg, creg)
+
+    circuit.x(qreg[0])
+    for i in order:
+        circuit.measure(qreg[i], creg[i])
+
+    folded = scaling.fold_gates_at_random(circuit, scale_factor=1.0)
+
+    assert get_counts(folded) == get_counts(circuit)
+
+
+def test_qiskit_measurement_order_is_preserved_two_registers():
+    """Tests measurement order is preserved when folding, i.e., the dictionary
+    of counts is the same as the original circuit on a noiseless simulator.
+    """
+    n = 4
+    qreg = qiskit.QuantumRegister(n)
+    creg1, creg2 = (
+        qiskit.ClassicalRegister(n // 2),
+        qiskit.ClassicalRegister(n // 2),
+    )
+    circuit = qiskit.QuantumCircuit(qreg, creg1, creg2)
+
+    circuit.x(qreg[0])
+    circuit.x(qreg[2])
+
+    # Some order of measurements.
+    circuit.measure(qreg[0], creg2[1])
+    circuit.measure(qreg[1], creg1[0])
+    circuit.measure(qreg[2], creg1[1])
+    circuit.measure(qreg[3], creg2[1])
+
+    folded = scaling.fold_gates_at_random(circuit, scale_factor=1.0)
+
+    assert get_counts(folded) == get_counts(circuit)
